[Ref] Fix cppcheck format string signed/unsigned mismatch warnings.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@12707 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/MIDIMacroDialog.cpp
@@ -172,7 +172,7 @@ void CMidiMacroSetup::UpdateMacroList(int macro)
 	for(int m = start; m <= end; m++)
 	{
 		// SFx
-		s.Format(_T("SF%X"), m);
+		s.Format(_T("SF%X"), static_cast<unsigned int>(m));
 		m_EditMacro[m].SetWindowText(s);
 
 		// Macro value:
