+++ mptrack/PatternFindReplaceDlg.cpp
@@ -512,7 +512,7 @@ void CFindReplaceTab::UpdateVolumeList()
 			m_cbnPCParam.InitStorage(ModCommand::maxColumnValue, 20);
 			for(int i = 0; i < ModCommand::maxColumnValue; i++)
 			{
-				wsprintf(s, _T("%02u: Parameter %02u"), i, i);
+				wsprintf(s, _T("%02u: Parameter %02u"), static_cast<unsigned int>(i), static_cast<unsigned int>(i));
 				m_cbnPCParam.SetItemData(m_cbnPCParam.AddString(s), i);
 			}
 		}
